requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
black==23.11.0
isort==5.12.0
pre-commit==3.6.0
